Split up rendering functions in submodule.
